alterando celular de long int para long long int, pq estava estourando e nao salvando)

+++ headers/colab.h
@@ -4,7 +4,7 @@
 typedef struct {
     int id;
     char nome[50];
-    long int celular;
+    long long int celular;
     char **servicosPrestados;
     int nServicos;
 }Colab;
